+++ src/lib/stores/resetGameStore.ts
@@ -1,0 +1,5 @@
+// /lib/stores/resetGameStore.ts
+
+import { writable } from "svelte/store";
+
+export const resetGameStore = writable<number>(0)